signer: move SSlibSigner and CryptoSigner
Move SSlibSigner and CryptoSigner to separate internal modules for a
cleaner code separation of signer interface and implementations.

This commit does not change functionality. SSlibSigner and CryptoSigner
(and its subclasses) remain publicly accessible via
`securesystemslib.signer`.
