+++ securesystemslib/signer/__init__.py
@@ -6,6 +6,12 @@ Some implementations are provided by default but more can be added by users.
 """
 from securesystemslib.signer._aws_signer import AWSSigner
 from securesystemslib.signer._azure_signer import AzureSigner
+from securesystemslib.signer._crypto_signer import (
+    CryptoSigner,
+    ECDSASigner,
+    Ed25519Signer,
+    RSASigner,
+)
 from securesystemslib.signer._gcp_signer import GCPSigner
 from securesystemslib.signer._gpg_signer import GPGKey, GPGSigner
 from securesystemslib.signer._hsm_signer import HSMSigner
@@ -13,12 +19,8 @@ from securesystemslib.signer._key import KEY_FOR_TYPE_AND_SCHEME, Key, SSlibKey
 from securesystemslib.signer._signature import Signature
 from securesystemslib.signer._signer import (
     SIGNER_FOR_URI_SCHEME,
-    ECDSASigner,
-    Ed25519Signer,
-    RSASigner,
     SecretsHandler,
     Signer,
-    SSlibSigner,
 )
 from securesystemslib.signer._sigstore_signer import SigstoreKey, SigstoreSigner
 from securesystemslib.signer._spx_signer import (
@@ -26,6 +28,7 @@ from securesystemslib.signer._spx_signer import (
     SpxSigner,
     generate_spx_key_pair,
 )
+from securesystemslib.signer._sslib_signer import SSlibSigner
 
 # Register supported private key uri schemes and the Signers implementing them
 SIGNER_FOR_URI_SCHEME.update(
